Create views folder
As setting up jest added a lot of config files in the root of the directory, it is a bit hard to see the application views - so moving them out into a dedicated directory.

Also renaming `init.js` to `menu.js` as this file is only responsible for setting up the menu.