Replace nunjuck link template with html
I created this template to reduce overhead of open in new tab attributes and text in external links, but I wasn't entirely happy with it as spacing and alignment is sometimes a bit off in the final html document.

I've opted to remove links opening in new tabs, as this isn't essential and does against the WCAG recommendation for links. There are a few articles recommending against this deciding this for users too, for example https://css-tricks.com/use-target_blank/ and https://www.smashingmagazine.com/2008/07/should-links-open-in-new-windows/

To replace my custom links templates with an <a> element I had some fun finding the right regex that I could use with Visual Studio Code's Find and Replace. It was a bit tricky to get the ignored () in the function working alongside capture groups, but I could use a regex to at least find the link text link and class names and place them into the right link structure:

```
// Find
"([a-zA-Z0-9'\s]+)",\s("[a-zA-Z0-9:\/.-]+")

// Replace
<a class="ld-link" href=$2>$1</a>
```