Update material icons import
Previously the entire material icons font face was being imported in the `@lookupdaily/styles` library, which was unnecessary as it was no longer being used in any of the styles.

Now directly importing via a stylesheet link, and only importing the icon(s) being used in the site. This improves the performance rating in Lighthouse - as it removes a blocking resource.